chore: rename React components to PascalCase
